+++ pybraid/src/lib.rs
@@ -3,7 +3,7 @@ mod utils;
 use std::collections::HashMap;
 use std::path::PathBuf;
 
-use braid::{EngineUpdateConfig, OracleT, PredictUncertaintyType};
+use braid::{EngineUpdateConfig, OracleT, HasStates, PredictUncertaintyType};
 use numpy::{IntoPyArray, PyArray1, ToPyArray};
 use pyo3::prelude::*;
 use pyo3::types::{PyDict, PyList};
@@ -21,6 +21,40 @@ struct Engine {
     rng: Xoshiro256Plus,
 }
 
+#[pyclass]
+struct ColumnMaximumLogpCache(
+    braid::ColumnMaximumLogpCache
+);
+
+#[pymethods]
+impl ColumnMaximumLogpCache {
+    #[staticmethod]
+    fn from_oracle(
+        engine: &Engine,
+        columns: &PyList,
+        given: Option<&PyDict>,
+    ) -> Self {
+        let col_ixs = column_indices(columns, &engine.col_indexer);
+
+        let given = dict_to_given(
+            given,
+            &engine.engine,
+            &engine.col_indexer,
+            &engine.value_maps,
+        );
+
+        let cache = braid::ColumnMaximumLogpCache::from_oracle(
+            &engine.engine,
+            &col_ixs,
+            &given,
+            None,
+        );
+
+        Self(cache)
+    }
+
+}
+
 #[pymethods]
 impl Engine {
     /// Create a Engine from metadata
@@ -34,6 +68,10 @@ impl Engine {
             rng: Xoshiro256Plus::from_entropy(),
             engine,
         }
+    }
+
+    fn n_states(&self) -> usize {
+        self.engine.n_states()
     }
 
     fn shape(&self) -> (usize, usize) {
@@ -174,8 +212,6 @@ impl Engine {
     ///     columns can either be indices (int) or names (str)
     /// n: int, optional
     ///     The number of records to simulate (default: 1)
-    /// state_ixs_opt: list(int), optional
-    ///     If specified, only the provided states will be used to simulate
     ///
     /// Returns
     /// -------
@@ -187,7 +223,6 @@ impl Engine {
         cols: &PyList,
         given: Option<&PyDict>,
         n: usize,
-        state_ixs_opt: Option<Vec<usize>>,
     ) -> Py<PyAny> {
         let col_ixs = column_indices(cols, &self.col_indexer);
         let given = dict_to_given(
@@ -199,7 +234,7 @@ impl Engine {
 
         let mut data = self
             .engine
-            .simulate(&col_ixs, &given, n, state_ixs_opt, &mut self.rng)
+            .simulate(&col_ixs, &given, n, None, &mut self.rng)
             .unwrap();
 
         Python::with_gil(|py| {
@@ -228,9 +263,6 @@ impl Engine {
     /// given: dict, optional
     ///     Column -> Value dictionary describing observations. Note that
     ///     columns can either be indices (int) or names (str)
-    /// state_ixs_opt: list(int), optional
-    ///     If specified, only the provided states will be used to compute the
-    ///     log liklihood, otherwise all states will be used (default).
     ///
     /// Returns
     /// -------
@@ -258,7 +290,6 @@ impl Engine {
         py: Python<'py>,
         values: &PyAny,
         given: Option<&PyDict>,
-        state_ixs_opt: Option<Vec<usize>>,
     ) -> &'py PyArray1<f64> {
         let (col_ixs, _, data) = pandas_to_logp_values(
             values,
@@ -276,31 +307,10 @@ impl Engine {
 
         let logps = self
             .engine
-            .logp(&col_ixs, &data, &given, state_ixs_opt)
+            .logp(&col_ixs, &data, &given, None)
             .unwrap();
 
         logps.into_pyarray(py)
-    }
-
-    fn col_max_logps(&self, cols: &PyList, given: Option<&PyDict>) -> Vec<f64> {
-        let col_ixs = utils::column_indices(cols, &self.col_indexer);
-
-        let given = dict_to_given(
-            given,
-            &self.engine,
-            &self.col_indexer,
-            &self.value_maps,
-        );
-
-        col_ixs
-            .iter()
-            .map(|&col_ix| {
-                let x = self.engine.predict(col_ix, &given, None).unwrap().0;
-                self.engine
-                    .logp(&[col_ix], &[vec![x]], &given, None)
-                    .unwrap()[0]
-            })
-            .collect()
     }
 
     fn logp_scaled<'py>(
@@ -308,8 +318,7 @@ impl Engine {
         py: Python<'py>,
         values: &PyAny,
         given: Option<&PyDict>,
-        state_ixs_opt: Option<Vec<usize>>,
-        col_max_logps: Option<Vec<f64>>,
+        col_max_logps: Option<&ColumnMaximumLogpCache>,
     ) -> &'py PyArray1<f64> {
         let (col_ixs, _, data) = pandas_to_logp_values(
             values,
@@ -329,8 +338,8 @@ impl Engine {
             &col_ixs,
             &data,
             &given,
-            state_ixs_opt,
-            col_max_logps.as_ref(),
+            None, 
+            col_max_logps.map(|cache| &cache.0),
         );
 
         logps.into_pyarray(py)
@@ -372,14 +381,14 @@ impl Engine {
         if with_uncertainty {
             let unc_type_opt = Some(PredictUncertaintyType::JsDivergence);
             let (pred, unc) =
-                self.engine.predict(col_ix, &given, unc_type_opt).unwrap();
+                self.engine.predict(col_ix, &given, unc_type_opt, None).unwrap();
             let value = datum_to_value(pred, col_ix, &self.engine.codebook);
             Python::with_gil(|py| {
                 let unc = unc.into_py(py);
                 (value, unc).into_py(py)
             })
         } else {
-            let (pred, _) = self.engine.predict(col_ix, &given, None).unwrap();
+            let (pred, _) = self.engine.predict(col_ix, &given, None, None).unwrap();
             datum_to_value(pred, col_ix, &self.engine.codebook)
         }
     }
@@ -523,6 +532,7 @@ impl Engine {
 #[pymodule]
 #[pyo3(name = "pybraid")]
 fn pybraid(_py: Python, m: &PyModule) -> PyResult<()> {
+    m.add_class::<ColumnMaximumLogpCache>()?;
     m.add_class::<Engine>()?;
     Ok(())
 }